Transactions history
- moved card to components/cards
- improved styled-components usage
